requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.25
langchain-core>=0.1.0
langchain-openai>=0.0.5

# LLM providers (通过langchain_community调用，不需要额外SDK)
openai>=1.0.0      # 备选方案

# Data handling and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File processing
python-markdown>=3.5.0
markdown-extensions>=0.1.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0
asyncio>=3.4.3

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 